requirements.txt
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==3.0.3
mss==4.0.2
numpy==1.16.2
opencv-python==4.0.0.21
Pillow==5.4.1
pyparsing==2.3.1
pytesseract==0.2.6
python-dateutil==2.8.0
six==1.12.0
tqdm
click
black
regex
paddle
paddleocr
